examples/requirements.txt
flaskrestframework